Reworkgraphics

+++ src/constants.rs
@@ -1,5 +1,5 @@
 // Number of tiles per edge = WINDOW_RES / (TILE_RES * ZOOM))
-pub(crate) const GAP_BETWEEN_TILES: u32 = 1;
+pub(crate) const GAP_BETWEEN_TILES: f32 = 1.0;
 
 pub(crate) const TILE_RES: f32 = 16.0;
 pub(crate) const ZOOM: f32 = 4.0; // This must be a power of 2
